add field macro for AppliedLens
