+++ macro/src/main/scala/monocle/macros/syntax/lens.scala
@@ -1,3 +1,3 @@
 package monocle.macros.syntax
 
-object lens extends GenApplyLensSyntax
+object lens extends GenFieldsLensSyntax
